Refactored to use longs instead of ints

+++ src/java/test/dev/scothammer/advent/day1/Day1Test.java
@@ -9,14 +9,14 @@ public class Day1Test {
 
     @Test
     public void givenTwoEmptyLists_sumOfDiffsIsZero() {
-        int[] list1 = {};
-        int[] list2 = {};
-        int sum = Day1.sumOfDiffs(list1, list2);
+        long[] list1 = {};
+        long[] list2 = {};
+        long sum = Day1.sumOfDiffs(list1, list2);
         assertEquals(0, sum);
     }
 
     private static class Day1 {
-        public static int sumOfDiffs(int[] list1, int[] list2) {
+        public static int sumOfDiffs(long[] list1, long[] list2) {
             return 0;
         }
     }
